requirements.txt
torch==2.1.2 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cu118
diffusers==0.30.1
transformers==4.44.2
opencv-python
safetensors
pillow==10.2.0
imageio
einops
peft
gradio==3.50.2
protobuf==3.20
accelerate
wandb
pytorch-lightning
opencv-contrib-python
omegaconf
open-clip-torch
jsonlines
diff-match-patch
git+https://github.com/openai/CLIP.git
deepspeed
tqdm
moviepy
imageio-ffmpeg
